# imports/revo/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Revo Treasury testnet simulation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /treasury/wallet/deposits:
    post:
      operationId: claimTreasuryDeposit
      tags: [treasury]
      summary: Credit a confirmed on-chain testnet USDC transfer to the treasury (verified against the Arc Testnet RPC)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositClaim"
      responses:
        "201":
          description: Deposit verified on-chain and credited to the treasury
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnchainTransfer"
  /treasury/wallet/withdrawals:
    post:
      operationId: requestTreasuryWithdrawal
      tags: [treasury]
      summary: Withdraw previously deposited testnet USDC back to the depositor's wallet (sent on-chain by the treasury)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalRequest"
      responses:
        "201":
          description: Withdrawal sent and confirmed on-chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnchainTransfer"
  /treasury/wallet/{address}/position:
    get:
      operationId: getTreasuryWalletPosition
      tags: [treasury]
      summary: Get a depositor wallet's net deposited balance and on-chain transfer history
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Depositor position and transfer ledger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletPosition"
...
components:
  schemas:
...
    OnchainTransfer:
      type: object
      required: [id, direction, wallet, amountUsdc, txHash, status, explorerTxUrl, createdAt, confirmedAt]
      properties:
        id:
          type: string
        direction:
          type: string
          description: deposit | withdrawal
        wallet:
          type: string
          description: Counterparty user wallet address (lowercased)
        amountUsdc:
          type: number
        txHash:
          type: ["string", "null"]
        status:
          type: string
          description: confirmed | pending | failed
        explorerTxUrl:
          type: ["string", "null"]
          description: Block explorer link for the transaction
        createdAt:
          type: string
        confirmedAt:
          type: ["string", "null"]
    WalletPosition:
      type: object
      required: [address, netDepositedUsdc, transfers]
      properties:
        address:
          type: string
        netDepositedUsdc:
          type: number
          description: Confirmed deposits minus confirmed/pending withdrawals - the amount this wallet can withdraw
        transfers:
          type: array
          items:
            $ref: "#/components/schemas/OnchainTransfer"
    DepositClaim:
      type: object
      required: [txHash]
      properties:
        txHash:
          type: string
          pattern: "^0x[0-9a-fA-F]{64}$"
          description: Hash of the on-chain testnet USDC transfer to the treasury address
    WithdrawalRequest:
      type: object
      required: [address, amountUsdc, issuedAt, signature]
      properties:
        address:
          type: string
          pattern: "^0x[0-9a-fA-F]{40}$"
          description: Destination wallet - must have a sufficient net deposited balance
        amountUsdc:
          type: number
          exclusiveMinimum: 0
          maximum: 1000000000
        issuedAt:
          type: string
          description: ISO timestamp the authorization message was signed at (valid for 5 minutes)
        signature:
          type: string
          pattern: "^0x[0-9a-fA-F]{130}$"
          description: EIP-191 personal_sign signature of the withdrawal authorization message, proving ownership of the destination wallet
...